--- app/views/layout.erb
@@ -1,167 +1,0 @@
-<html>
-  <head>
-    <title>OpenPolitics Votebot</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css">
-    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css">
-    <link href="http://openpolitics.org.uk/css/main.css" rel="stylesheet">
-    <link href="/stylesheets/custom.css" rel="stylesheet">
-    <link type="image/png" sizes="64x64" href="http://openpolitics.org.uk/images/logo64.png" rel="icon"></link>
-    <style>
-       body {
-         padding-top: 50px;
-       }
-     </style>
-  </head>
-  <body>
-    
-    <div class="navbar navbar-inverse navbar-fixed-top">
-        <div class="container">
-          <div class="navbar-header">
-            <button type="button" class="navbar-toggle" data-toggle="collapse" data-target=".navbar-collapse">
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-            </button>
-            <a class="navbar-brand" href="http://openpolitics.org.uk/index.html">
-              <img src="http://openpolitics.org.uk/images/logo64_inverted.png" alt="logo" height="20" style="vertical-align: text-bottom">&nbsp;OpenPolitics
-            </a>
-          </div>
-          <div class="collapse navbar-collapse">
-            <ul class="nav navbar-nav">
-              <li class="dropdown">
-                <a class="dropdown-toggle" data-toggle="dropdown" href="http://openpolitics.org.uk/manifesto/index.html">Manifesto <span class="caret"></span></a>
-                <ul class="dropdown-menu">
-                  <li> <a href="http://openpolitics.org.uk/manifesto/index.html">Introduction</a></li>
-                  
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/communications.html">Communications</a></li>
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/crime.html">Crime &amp; Justice</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/culture.html">Culture &amp; Media</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/democracy.html">Democracy</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/economy.html">Economy</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/education.html">Education</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/energy.html">Energy</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/environment.html">Environment</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/foreign_policy.html">Foreign Policy</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/health.html">Health</a></li>
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/constituencies/horsham.html">Horsham</a></li>
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/housing.html">Housing</a></li>
-                    
-                  
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/infrastructure.html">Infrastructure</a></li>
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/local_government.html">Local Government</a></li>
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/military.html">Military</a></li>
-                    
-                  
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/science.html">Science</a></li>
-                    
-                  
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/society.html">Society</a></li>
-                    
-                  
-                    
-                  
-                    
-                  
-                    
-                      <li> <a href="http://openpolitics.org.uk/manifesto/transport.html">Transport</a></li>
-                    
-                  
-                </ul>
-              </li>
-              <li><a href="http://votebot.openpolitics.org.uk">Votes</a></li>
-              <li><a href="http://openpolitics.org.uk/candidates.html">Candidates</a></li>
-              <li><a href="http://openpolitics.org.uk/contributing.html">How to Contribute</a></li>
-            </ul>
-            
-            <!--<form class="navbar-form navbar-right" role="search" action='/search.html' method='get' style='margin-right: 20px'>
-              <div class="form-group">
-                <input type="text" class="form-control" name='query' placeholder="Search">
-              </div>
-            </form>-->
-          </div><!--/.nav-collapse -->
-        </div>
-      </div>  
-    
-    <div class='container'>
-
-      <%= yield %>
-
-      <footer class='well'>
-        Code available on <a href='http://github.com/openpolitics/votebot'>GitHub</a>. Error tracking generously provided by <a href='http://bugsnag.com'>Bugsnag</a>.
-      </footer>
-      
-    </div>
-    
-    <script src="//code.jquery.com/jquery-2.1.3.min.js"></script>
-    <script src="//maxcdn.bootstrapcdn.com/bootstrap/3.3.2/js/bootstrap.min.js"></script>
-
-  </body>
-</html>
