+++ app/views/shared/_person.html.erb
